CARDS-1337: Enable patient authentication Make the visit ACL restriction more generic, it should not mention ePRO-specific visits, it should instead apply to any kind of subject. Also move it in the cards-permissions module since it is generic enough.
